requirements.txt
numpy
opencv-python
maturin
Pillow
